refactored routes and auth into correct folders
