+++ server/auth/auth.test.js
@@ -2,7 +2,7 @@ const request = require('supertest-as-promised')
 const {expect} = require('chai')
 const db = require('APP/db')
 const User = require('APP/db/models/user')
-const app = require('./start')
+const app = require('../start')
 
 const alice = {
   username: 'alice@secrets.org',
@@ -35,24 +35,24 @@ describe('/api/auth', () => {
         .post('/api/auth/local/login')
         .send({username: alice.username, password: 'wrong'})
         .expect(401)
-      )      
+      )
   })
 
   describe('GET /whoami', () => {
     describe('when logged in,', () => {
       const agent = request.agent(app)
       before('log in', () => agent
-        .post('/api/auth/local/login') 
+        .post('/api/auth/local/login')
         .send(alice))
 
       it('responds with the currently logged in user', () =>
         agent.get('/api/auth/whoami')
-          .set('Accept', 'application/json')        
-          .expect(200)          
+          .set('Accept', 'application/json')
+          .expect(200)
           .then(res => expect(res.body).to.contain({
             email: alice.username
           }))
-      )      
+      )
     })
 
     it('when not logged in, responds with an empty object', () =>
@@ -66,7 +66,7 @@ describe('/api/auth', () => {
     const agent = request.agent(app)
 
     before('log in', () => agent
-      .post('/api/auth/local/login') 
+      .post('/api/auth/local/login')
       .send(alice))
 
     it('logs you out and redirects to whoami', () => agent
