+++ server/routes/users.test.js
@@ -2,7 +2,7 @@ const request = require('supertest-as-promised')
 const {expect} = require('chai')
 const db = require('APP/db')
 const User = require('APP/db/models/user')
-const app = require('./start')
+const app = require('../start')
 
 describe('/api/users', () => {
   describe('when not logged in', () => {
@@ -10,7 +10,7 @@ describe('/api/users', () => {
       request(app)
         .get(`/api/users/1`)
         .expect(401)
-    )    
+    )
 
     it('POST creates a user', () =>
       request(app)
@@ -32,7 +32,7 @@ describe('/api/users', () => {
         .redirects(1)
         .then(res => expect(res.body).to.contain({
           email: 'eve@interloper.com'
-        }))        
+        }))
     )
   })
 })
